requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-rest-auth
Django
django-allauth
django-environ
djangorestframework
idna
mercadopago
oauthlib
Pillow
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
sqlparse
tzdata
urllib3==1.26.15